Orcard Exp file utility
This is a tool I created to simply my life a bit wrt Orcad.

It allows one to work with exp file created by orcad. It can
be used to check out differences between revisions or to update
certain properties from a predetermined database (a csv file)
of values and so. Check the readme for bit more details.

This is the version which I have found from around this time.
If I find older revisions, I will be commiting them later.